requirements.txt
urllib3 >= 1.26.18, < 2.0

selenium >= 3.141.0, < 4.0
Appium-Python-Client == 1.3.0

behave == 1.2.6
PyHamcrest >= 2.0.0, < 2.1.0
envparse < 0.3
requests >= 2.31.0, < 3.0.0
babel == 2.9.1
py-postgresql == 1.2.2
colorlog >= 6.6.0, < 7.0.0
validators == 0.18.2
mimesis == 4.1.3
nested-lookup == 0.2.23
Pillow >= 10.1.0, < 11.0.0
retrying == 1.3.3
tenacity >= 8.0.1, < 9.0.0
pyderman == 3.3.2
xmltodict == 0.12.0